# http/output/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"IPet": {
				"properties": {
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"species": {
						"type": "string"
					},
					"customerId": {
						"type": "string"
					},
					"petId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"age",
					"species",
					"customerId",
					"petId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IPet.Exclude_keyofIPet.petId-or-customerId__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"species": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"age",
					"species"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"AddPetsDTO": {
				"properties": {
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"species": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"age",
					"species"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "example-server-tsoa",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/pets": {
			"get": {
				"operationId": "GetPets",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IPet"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AddPets",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPet"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddPetsDTO"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}